requirements.txt
asgiref==3.5.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-ckeditor==6.4.0
django-colorfield==0.7.0
django-dotenv==1.4.2
django-import-export==2.8.0
django-js-asset==2.0.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
pycodestyle==2.8.0
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
xlrd==2.0.1
xlwt==1.3.0
